refactor(nvlink-manager): extract NVLink code and config
- move NvlPartitionMonitor and NMX-M client pool into new crate
- move NvLinkConfig and DEFAULT_NMX_M_NAME out of api
- add crate-local NvLinkManagerError/Result
- move TransactionVending to api-db
