+++ crates/api/src/lib.rs
@@ -58,8 +58,6 @@ mod machine_validation;
 mod measured_boot;
 mod mqtt_state_change_hook;
 mod network_segment;
-mod nvl_partition_monitor;
-mod nvlink;
 mod rack;
 mod redfish;
 mod run;
