+++ crates/nvlink-manager/src/errors.rs
@@ -1,0 +1,38 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use db::DatabaseError;
+use rpc::errors::RpcDataConversionError;
+
+#[derive(thiserror::Error, Debug)]
+pub enum NvLinkManagerError {
+    #[error("Database error: {0}")]
+    DatabaseError(#[from] DatabaseError),
+    #[error("Can not convert between RPC data model and internal data model - {0}")]
+    RpcDataConversionError(#[from] RpcDataConversionError),
+    #[error("Internal error: {message}")]
+    Internal { message: String },
+}
+
+impl NvLinkManagerError {
+    /// Creates a `Internal` error with the given error message
+    pub fn internal(message: String) -> Self {
+        Self::Internal { message }
+    }
+}
+
+pub type NvLinkManagerResult<T> = Result<T, NvLinkManagerError>;
